app/requirements.txt
Flask
Flask-WTF
pymongo
openai
boto3
Werkzeug
python-dotenv
Flask-Cors
moviepy

